% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_tukey_tests.R
\name{anova_tukey_tests}
\alias{anova_tukey_tests}
\title{anova_tukey_tests}
\usage{
anova_tukey_tests(data)
}
\arguments{
\item{data}{Conjunto de datos a utilizar}
}
\value{
una tabla con el analisis de las pruebas ANOVA y/o Tukey en los datos
}
\description{
Esta funciòn permite realizar pruebas ANOVA y Tukey en los datos,
toma un conjunto de datos como argumento y  permite especificar qué análisis deseas realizar.
}
\examples{
 expr <- data a utilizar
anova_tukey_tests(expr, 1) Para realizar el Test ANOVA
anova_tukey_tests(expr, 2) para realizar el Test Tukey
}
