% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_expression.R
\name{boxplot_expression}
\alias{boxplot_expression}
\title{boxplot_expression}
\usage{
boxplot_expression(data, x_var, y_var)
}
\arguments{
\item{data}{Conjunto de datos a utilizar para generar el gráfico}

\item{x_var}{una variable para el eje x (x_var), dataset de la data, para el caso "tipos de cancer"}

\item{y_var}{una variable para el eje y (y_var), nombre de Gen}
}
\value{
un boxplot de la expresion de genes en diferentes tipos de cancer
}
\description{
Esta función genera un gráfico de caja para visualizar la distribución de
los valores de las variables del eje y para cada valor de la variable del eje x.
Si se especifica el valor "all" para el argumento y_var, la función incluirá todas las columnas
del conjunto de datos que no sean la variable del eje x ni la columna "bcr_patient_barcode".
}
\examples{
 expr <- data a utilizar
"dataset" <- Tipos de cancer
XBP1 = Gen
boxplot_expression(expr, "dataset", "XBP1") Genera un gráfico de caja para un solo gen
boxplot_expression(expr, "dataset", "all") Genera un gráfico de caja para todos los genes disponibles en el conjunto de datos

}
