% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jiterplot_expression.R
\name{jiterplot_expression}
\alias{jiterplot_expression}
\title{jiterplot_expression}
\usage{
jiterplot_expression(data, x_var, y_vars)
}
\arguments{
\item{data}{Conjunto de datos a utilizar para generar el gráfico}

\item{x_var}{una variable para el eje x (x_var), dataset de la data, para el caso "tipos de cancer"}

\item{y_var}{una variable para el eje y (y_var), nombre de Gen}
}
\value{
un boxplot con puntos de disperion, de la expresion de genes en diferentes tipos de cancer
}
\description{
Esta función genera un gráfico de caja con puntos jitter para
visualizar la distribución de los valores de las variables del eje
y para cada valor de la variable del eje x.
El gráfico incluye facetas para cada variable del eje y.
si se especifica el valor "all" para el argumento y_vars,
la función incluirá todas las columnas del conjunto de datos que
no sean la variable del eje x ni la columna "bcr_patient_barcode".
La función verifica si las variables especificadas para los ejes x
e y se encuentran en el conjunto de datos
}
\examples{
 expr <- data a utilizar
"dataset" <- Tipos de cancer
XBP1 = Gen
jiterplot_expression(expr, "dataset", "XBP1") Genera un gráfico de caja para un solo gen
jiterplot_expression(expr, "dataset", "all") Genera un gráfico de caja para todos los genes disponibles en el conjunto de datos
}
