% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_pca}
\alias{plot_pca}
\title{plot_pca}
\usage{
plot_pca(data)
}
\arguments{
\item{data}{Conjunto de datos a utilizar para generar el gr치fico}
}
\value{
El gr치fico PCA de dos dimensiones.
}
\description{
Esta funci칩n toma un conjunto de datos con
variables cuantitativas y genera un gr치fico PCA de dos dimensiones.
}
\examples{
expr <- data a utilizar
plot_pca(expr)

}
