% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdensity_cancer.R
\name{plotdensity_cancer}
\alias{plotdensity_cancer}
\title{plotdensity_cancer}
\usage{
plotdensity_cancer(data, dataset_col, cancer_type)
}
\arguments{
\item{data}{un objeto data.frame que contiene la expresión data}

\item{dataset_col}{una cadena que indica el nombre de la columna que contiene la información del conjunto de datos}

\item{cancer_type}{una cadena que indica el tipo de cáncer a trazar o “all” para trazar todos los tipos de cáncer}
}
\value{
un gráfico de densidad para los datos de expresión del tipo de cáncer especificado
}
\description{
Esta función genera un gráfico de densidad para visualizar la distribución de los valores de expresión génica para un tipo de cáncer específico
o para todos los tipos de cáncer disponibles en el conjunto de datos.
}
\examples{

expr <- data a utilizar
"dataset" <- Tipos de cancer
XBP1 = Gen
plotdensity_cancer(expr, "dataset",  "BRCA") Genera un gráfico de caja para un solo gen
plotdensity_cancer(expr, "dataset", "all") Genera un gráfico de caja para todos los genes disponibles en el conjunto de datos
}
